refactor: extract OpenAI route-model transport
